% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adEnv.R
\name{ad.variant}
\alias{ad.variant}
\title{Create Variant Functions (more) Suitable for AD.}
\usage{
ad.variant(f)
}
\arguments{
\item{f}{the original function.}
}
\description{
\code{ad.variant} is a helper function which creates variant functions more suitable for automatic differentiation.
}
