% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autodiff.R
\name{ad_setup}
\alias{ad_setup}
\title{Do initial setup for package autodiffr.}
\usage{
ad_setup(...)
}
\arguments{
\item{...}{arguments passed to \code{JuliaCall::julia_setup}.}
}
\description{
\code{ad_setup} does the initial setup for package autodiffr.
}
\examples{
\dontrun{
ad_setup()
}

}
