% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adEnv.R
\name{ad_variant}
\alias{ad_variant}
\title{Create Variant Functions (more) Suitable for AD.}
\usage{
ad_variant(f)
}
\arguments{
\item{f}{the original function.}
}
\description{
\code{ad_variant} is a helper function which creates variant functions more suitable for automatic differentiation.
}
