% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zmakeinterface.R
\name{autodifffunc}
\alias{autodifffunc}
\alias{makeJacobianFunc}
\alias{makeHessianFunc}
\alias{makeGradFunc}
\alias{makeDerivFunc}
\title{Create (Optimized) Gradient, Jacobian and Hessian Functions using Automatic Differentiation.}
\usage{
makeJacobianFunc(func, ..., mode = c("reverse", "forward"), x = NULL,
  chunk_size = NULL, use_tape = FALSE, compiled = FALSE, debug = TRUE)

makeHessianFunc(func, ..., mode = c("reverse", "forward"), x = NULL,
  chunk_size = NULL, use_tape = FALSE, compiled = FALSE, debug = TRUE)

makeGradFunc(func, ..., mode = c("reverse", "forward"), x = NULL,
  chunk_size = NULL, use_tape = FALSE, compiled = FALSE, debug = TRUE)

makeDerivFunc(func, ..., mode = c("reverse", "forward"), x = NULL,
  chunk_size = NULL, use_tape = FALSE, compiled = FALSE, debug = TRUE)
}
\arguments{
\item{func}{the target function to calculate gradient, jacobian or hessian.}

\item{...}{other arguments passed to the target function \code{func}.}

\item{mode}{whether to use forward or reverse mode automatic differentiation.}

\item{x}{If \code{x} is given, then the returning functiopn
will be optimized w.r.t. the input(s) of the same shape with \code{x}.
Note that if \code{x} is not given, then all the following parameters will be ignored.}

\item{chunk_size}{the chunk size to use in forward mode automatic differentiation.
Note that this parameter is only effective when \code{mode = "forward"}.}

\item{use_tape}{whether or not to use tape for reverse mode automatic differentiation.
Note that although \code{use_tape} can greatly improve the performance sometime,
it can only calculate the derivatives w.r.t a certain branching,
so it may not give the correct result for functions containing things like \code{if}.
And this parameter is only effective when \code{mode = "reverse"}.}

\item{compiled}{whether or not to use compiled tape for reverse mode automatic differentiation.
Note that tape compiling can take a lot of time and
this parameter is only effective when \code{use_tape = TRUE}.}

\item{debug}{Whether to activate debug mode.
With the debug mode, users can have more informative error messages.
Without the debug mode, functions will be more performant.}
}
\value{
if \code{x} is not given, the returned function will be a general function to calculate derivatives,
which accepts other arguments to \code{func} besides \code{x};
if \code{x} is given, the returned function will be an optimized function speciallised for inputs of the same
shape with the given \code{x}.
}
\description{
These functions give (optimized) gradient, jacobian or hessian functions for the target function.
}
