% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colSums.R
\name{colSums}
\alias{colSums}
\alias{rSums}
\alias{rMeans}
\alias{cSums}
\alias{cMeans}
\title{Row and Column Sums and Means Compatable with JuliaObject}
\usage{
rSums(x)

rMeans(x)

cSums(x)

cMeans(x)
}
\arguments{
\item{x}{\code{R} or \code{julia} numeric matrices or vectors.}
}
\description{
Calculate row and column sums and means compatible with \code{JuliaObject}.
}
