% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagm.R
\name{diagm}
\alias{diagm}
\title{Construct a diagonal matrix}
\usage{
diagm(x)
}
\arguments{
\item{x}{a vector or 1D array.}
}
\value{
a matrix with `x` as diagonal and zero off-diagonal entries.
}
\description{
`diag` construct a diagnonal matrix.
}
