% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{grad}
\alias{grad}
\title{Calculate the gradient of a function.}
\usage{
grad(func, x = NULL)
}
\arguments{
\item{func}{the function you want to caculate the gradient,
it should be a function with a scalar result.}

\item{x}{the point where the gradient is calculated.
If not given (or NULL), return will be the gradient function.}
}
\description{
\code{grad} calculates the gradient of a function.
}
