% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matmult.R
\name{\%x\%}
\alias{\%x\%}
\title{Matrix Multiplication Compatable with JuliaObject}
\usage{
x \%x\% y
}
\arguments{
\item{x, y}{numeric or complex matrices or vectors.}
}
\description{
`%x%` is matrix multiplication operator compatible with `JuliaObject`.
}
