% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{Jmatrix}
\alias{Jmatrix}
\title{`matrix` Compatible with `JuliaObject`.}
\usage{
Jmatrix(data, nrow, ncol)
}
\arguments{
\item{data}{the data vector used to create matrix.}

\item{nrow}{number of rows of the matrix.}

\item{ncol}{number of columns of the matrix.}
}
\description{
`Jmatrix` creates a matrix from the given set of values.
}
