% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia_array.R
\name{JuliaArray}
\alias{JuliaArray}
\title{Create Julia Arrays}
\usage{
JuliaArray(data, n1, n2, ...)
}
\arguments{
\item{data}{the data to create the array.}

\item{n1, n2, ...}{the dimensions.}
}
\description{
\code{JuliaArray} is a helper function to create Julia arrays including vectors, matrices and etc.
And it can be used in functions to facilitate automatic differentiation.
}
\details{
If n1 is missing, then all the dimensions will be ignored,
and the result will be a Julia array coresponding to the data argument.
If n1, n2, ... are given, they will determine the shape of the result,
and data will be used for the content.
}
\examples{

\donttest{ ## setup is quite time consuming
  JuliaArray(0, 3) ## will return a Julia vector with length 3 and filled with 0
  JuliaArray(0, 3, 4) ## will return a Julia matrix of 3x4 and filled with 0
  JuliaArray(matrix(1, 2, 2)) ## dimension is not given, will return a Julia matrix
  JuliaArray(matrix(1, 2, 2), 4) ## dimension is given, will return a Julia vector
}

}
