% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Apply a Function to Multiple List or Vector Arguments}
\usage{
map(FUN, ..., MoreArgs = NULL, SIMPLIFY = TRUE, USE.NAMES = TRUE)
}
\arguments{
\item{FUN}{function to apply.}

\item{...}{arguments.}

\item{MoreArgs}{a list of other arguments to FUN.}

\item{SIMPLIFY}{whether or not to simplify the result.}

\item{USE.NAMES}{whether or not to use names.}
}
\description{
`map` is same as `mapply`, except more efficient for `JuliaObject`.
}
