% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros_ones.R
\name{ones}
\alias{ones}
\title{Create an array of all ones with the same element type and dims as \code{x}.}
\usage{
ones(x)
}
\arguments{
\item{x}{\code{R} or \code{julia} matrices or vectors.}
}
\description{
\code{ones} create an array of all ones with the same element type and dims as \code{x}.
}
