% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.autodiffr_fit}
\alias{print.autodiffr_fit}
\title{Print method for autodiffr_fit objects}
\usage{
\method{print}{autodiffr_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{autodiffr_fit}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print method for autodiffr_fit objects
}
