% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_gradcheck.R
\name{print.autodiffr_gradcheck}
\alias{print.autodiffr_gradcheck}
\title{Print method for autodiffr_gradcheck objects}
\usage{
\method{print}{autodiffr_gradcheck}(x, ...)
}
\arguments{
\item{x}{An object of class \code{autodiffr_gradcheck}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print method for autodiffr_gradcheck objects
}
