% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_vcov_info.R
\name{print.autodiffr_vcov_info}
\alias{print.autodiffr_vcov_info}
\title{Print method for autodiffr_vcov_info objects}
\usage{
\method{print}{autodiffr_vcov_info}(x, ...)
}
\arguments{
\item{x}{An object of class \code{autodiffr_vcov_info}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print method for autodiffr_vcov_info objects
}
