% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.autodiffr_fit}
\alias{print.summary.autodiffr_fit}
\title{Print method for summary.autodiffr_fit objects}
\usage{
\method{print}{summary.autodiffr_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.autodiffr_fit}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
Invisibly returns \code{x}
}
\description{
Print method for summary.autodiffr_fit objects
}
