% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.autodiffr_fit}
\alias{summary.autodiffr_fit}
\title{Summary method for autodiffr_fit objects}
\usage{
\method{summary}{autodiffr_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{autodiffr_fit}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
An object of class \code{summary.autodiffr_fit} with detailed information
}
\description{
Summary method for autodiffr_fit objects
}
