% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_forward_derivative}
\alias{test_forward_derivative}
\title{derivative of 1 + x + x*x + 1/x + log(x) in forward mode}
\usage{
test_forward_derivative()
}
\description{
derivative of 1 + x + x*x + 1/x + log(x) in forward mode
}
