% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_forward_gradient}
\alias{test_forward_gradient}
\title{gradient of (x * x.exp()).sum() in forward mode}
\usage{
test_forward_gradient()
}
\description{
gradient of (x * x.exp()).sum() in forward mode
}
