% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_reverse_derivative}
\alias{test_reverse_derivative}
\title{derivative of 1 + x + x*x + 1/x + log(x) in reserse mode}
\usage{
test_reverse_derivative()
}
\description{
derivative of 1 + x + x*x + 1/x + log(x) in reserse mode
}
