% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_reverse_derivative_multi}
\alias{test_reverse_derivative_multi}
\title{partial derivatives of 1 + x + y + z + x*y + y*z + x*z + x*y*z + exp(x/y + y/z) in reverse mode}
\usage{
test_reverse_derivative_multi()
}
\description{
partial derivatives of 1 + x + y + z + x*y + y*z + x*z + x*y*z + exp(x/y + y/z) in reverse mode
}
