% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_reverse_gradient}
\alias{test_reverse_gradient}
\title{gradient of (x * x.exp()).sum() in reverse mode}
\usage{
test_reverse_gradient()
}
\description{
gradient of (x * x.exp()).sum() in reverse mode
}
