% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiffTests.R
\docType{data}
\name{TESTING_FUNCS}
\alias{TESTING_FUNCS}
\title{Testing functions.}
\format{An object of class \code{list} of length 8.}
\usage{
TESTING_FUNCS
}
\description{
Lists of testing functions, which can be used for examples, testing and benchmarks.
}
\keyword{datasets}
