% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.R
\name{vcov.autodiffr_fit}
\alias{vcov.autodiffr_fit}
\title{Extract Variance-Covariance Matrix from autodiffr_fit}
\usage{
\method{vcov}{autodiffr_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{autodiffr_fit}}

\item{...}{Additional arguments (currently ignored)}
}
\value{
A variance-covariance matrix, or NULL if not available
}
\description{
Extract Variance-Covariance Matrix from autodiffr_fit
}
