% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zeros_ones.R
\name{zeros}
\alias{zeros}
\title{Create an array of all zeros with the same element type and dims as \code{x}.}
\usage{
zeros(x)
}
\arguments{
\item{x}{\code{R} or \code{julia} matrices or vectors.}
}
\description{
\code{zeros} create an array of all zeros with the same element type and dims as \code{x}.
}
