% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zinterface.R
\name{autodiff}
\alias{autodiff}
\alias{jacobian}
\alias{hessian}
\alias{grad}
\alias{deriv}
\title{Calculate Gradient, Jacobian and Hessian using Automatic Differentiation.}
\usage{
jacobian(func, x = NULL, ..., mode = c("reverse", "forward"), xsize = x,
  chunk_size = NULL, use_tape = FALSE, compiled = FALSE)

hessian(func, x = NULL, ..., mode = c("reverse", "forward"), xsize = x,
  chunk_size = NULL, use_tape = FALSE, compiled = FALSE)

grad(func, x = NULL, ..., mode = c("reverse", "forward"), xsize = x,
  chunk_size = NULL, use_tape = FALSE, compiled = FALSE)

deriv(func, x = NULL, ..., mode = c("reverse", "forward"), xsize = x,
  chunk_size = NULL, use_tape = FALSE, compiled = FALSE)
}
\arguments{
\item{func}{the target function to calculate gradient, jacobian or hessian.}

\item{x}{the input(s) where derivative is (are) taken.}

\item{...}{other arguments passed to the target function \code{func}.}

\item{mode}{whether to use forward or reverse mode automatic differentiation.}

\item{xsize}{If \code{x} is not given and \code{xsize} is given, then the returning functiopn
will be optimized w.r.t. the input(s) of the same size with \code{x}.
Note that if \code{xsize} is not given, then all the following parameters will be ignored.}

\item{chunk_size}{the chunk size to use in forward mode automatic differentiation.
Note that this parameter is only effective when \code{mode = "forward"}.}

\item{use_tape}{whether or not to use tape for reverse mode automatic differentiation.
Note that although \code{use_tape} can greatly improve the performance sometime,
it can only calculate the derivatives w.r.t a certain branching,
so it may not give the correct result for functions containing things like \code{if}.
And this parameter is only effective when \code{mode = "reverse"}.}

\item{compiled}{whether or not to use compiled tape for reverse mode automatic differentiation.
Note that tape compiling can take a lot of time and
this parameter is only effective when \code{use_tape = TRUE}.}
}
\value{
if \code{x} is given, then return will be derivatives;
if \code{x} is not given, then return will be a function to calculate derivatives.
}
\description{
These functions calculate gradient, jacobian or hessian for the target function.
}
