% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{ReverseDiff}
\alias{ReverseDiff}
\alias{reverse.grad}
\alias{reverse.jacobian}
\alias{reverse.hessian}
\alias{reverse.grad.config}
\alias{reverse.jacobian.config}
\alias{reverse.hessian.config}
\title{Wrapper functions for API of \code{ReverseDiff.jl}.}
\usage{
reverse.grad(f, input, cfg = JuliaCall::julia_call("GradientConfig", input))

reverse.jacobian(f, input, cfg = JuliaCall::julia_call("JacobianConfig",
  input))

reverse.hessian(f, input, cfg = JuliaCall::julia_call("HessianConfig", input))

reverse.grad.config(input, tp = JuliaCall::julia_call("Rawtape"))

reverse.jacobian.config(input, tp = JuliaCall::julia_call("Rawtape"))

reverse.hessian.config(input, gtp = JuliaCall::julia_call("Rawtape"),
  jtp = JuliaCall::julia_call("Rawtape"))
}
\arguments{
\item{f}{the function you want to calulate the gradient, jacobian and hessian.
Note that \code{f(x)} should be a scalar for \code{grad} and \code{hessian},
a vector of length greater than 1 for \code{jacobian}.}

\item{input}{the point where you take the gradient, jacobian and hessian.
Note that it should be a a vector of length greater than 1.
If you want to calulate the derivative of a function, you can considering using \code{forward.deriv}.}

\item{cfg}{Config objects which contains the preallocated tape and work buffers
used by reverse mode automatic differentiation.
\code{ReverseDiff}'s API methods will allocate the Config object automatically by default,
but you can preallocate them yourself and reuse them for subsequent calls to reeuce memory usage.}

\item{tp, gtp, jtp}{the tapes to construct the Config objects for \code{ReverseDiff},
which are for advantage usage.
See
\url{http://www.juliadiff.org/ReverseDiff.jl/api/#the-abstracttape-api}
for more details.}
}
\value{
\code{reverse.grad}, \code{reverse.jacobian} and \code{reverse.hessian} return
the gradient, jacobian and hessian of \code{f} correspondingly evaluated at \code{input}.
\code{reverse.grad.config}, \code{reverse.jacobian.config} and \code{reverse.hessian.config}
return Config instances containing the preallocated tape and work buffers used by
reverse mode automatic differentiation.
}
\description{
Wrapper functions for API of \code{ReverseDiff.jl} at
\url{http://www.juliadiff.org/ReverseDiff.jl/api/}.
These functions can help you calculate gradient, jacobian and hessian
for your functions using reverse mode automatic differentiation.
For more details, see \url{http://www.juliadiff.org/ReverseDiff.jl/api/}.
}
